풀이완료 Language_Coder/문자열2/형성평가

+++ Language_Coder/문자열1/형성평가/FormativeString107.cpp
@@ -24,7 +24,7 @@ void FormativeString107::Code()
 
 	std::getline(std::cin, str);
 
-	int count{ 2 };
+	size_t count{ 2 };
 
 	for (size_t i = 0; i < str.size(); i++)
 	{
